Create UserInput.cpp
In this program:

We declare three variables (age, height, and name) to store the user's input.
We use cout to prompt the user to enter their age, height, and name.
We use cin to read the user's input into the respective variables. We use cin.ignore() before reading the name to clear the input buffer so that we can read a line of text (including spaces).
Finally, we display the user's input using cout.